More robust mechanism for optionally collapsing furigana
The parser now returns an array of arrays if items (an array of groups)

So when we collapse, we collapse per-group, not per "chunk of continuous kanji that
has furigana".

Per group collapsing is more robust, and more accurately reflects the
intention of the input text.

It doesn't matter so much at this point since we're not using the
collapsing feature on the UI anyway